+++ src/collision_system/collider.rs
@@ -1,4 +1,6 @@
-use macroquad::math::Vec2;
+use macroquad::{color::{BLACK, RED}, math::Vec2, shapes::{draw_rectangle_ex, DrawRectangleParams}};
+
+use crate::renderer::artist::DrawCall;
 
 ///Each entity is either assigned a circle or rectangular collider.
 pub trait Collider{